requirements.txt
boto3==1.20.14
botocore==1.23.14
certifi==2021.10.8
charset-normalizer==2.0.8
feedparser==6.0.8
idna==3.3
jmespath==0.10.0
pocket-api==0.1.5
python-dateutil==2.8.2
requests==2.26.0
s3transfer==0.5.0
sgmllib3k==1.0.0
six==1.16.0
urllib3==1.26.7
